pyproject.toml
[project]
name = "switchboard-simulator"
version = "0.1.0"
description = "The Switchboard AI Game Simulator - a game of clandestine communication and deduction"
authors = [
    {name = "Jacob Matson", email = "jacob@example.com"},
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "types-pyyaml>=6.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
switchboard = "switchboard.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["switchboard"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]
